setup tool command
