+++ src/cli.rs
@@ -1,0 +1,12 @@
+use crate::action::{handle_action, Action};
+use dip::cli::CliPlugin;
+
+#[derive(CliPlugin, clap::Parser)]
+#[clap(version)]
+struct Cli {
+    #[clap(short, long)]
+    path: Option<String>,
+
+    #[clap(subcommand)]
+    action: Action,
+}
