+++ src/tool.rs
@@ -1,21 +1,46 @@
 use dip::{
-    bevy::{ecs::event::EventReader, log},
+    bevy::{
+        app::{App, Plugin},
+        ecs::event::EventReader,
+        log,
+    },
     cli::SubcommandPlugin,
 };
 
-#[derive(SubcommandPlugin, clap::Subcommand, Clone, Debug)]
-pub enum ToolAction {
-    Install,
-}
+pub struct ToolPlugin;
 
-impl Default for ToolAction {
-    fn default() -> Self {
-        Self::Install
+impl Plugin for ToolPlugin {
+    fn build(&self, app: &mut App) {
+        app.add_plugin(ToolActionPlugin)
+            .add_system(handle_list_tool)
+            .add_system(handle_add_tool);
     }
 }
 
-pub fn handle_tool_install(mut events: EventReader<InstallToolAction>) {
+#[derive(SubcommandPlugin, clap::Subcommand, Clone, Debug)]
+pub enum ToolAction {
+    List,
+    Add(AddAction),
+}
+
+#[derive(clap::Args, Clone, Debug)]
+pub struct AddAction {
+    name: String,
+}
+
+pub fn handle_list_tool(mut events: EventReader<ListToolAction>) {
     for e in events.iter() {
-        log::info!("{e:#?}");
+        log::info!("{e:?}");
+    }
+}
+
+pub fn handle_add_tool(mut events: EventReader<AddToolAction>) {
+    for e in events.iter() {
+        match e.name.as_str() {
+            "tailwind" => {
+                log::info!("{e:?}");
+            }
+            _ => {}
+        }
     }
 }
